refactor: Refactor transcribe samples into two separate samples (local-file vs gcs) and remove io dependency. Refactor tests.
