+++ samples/snippets/transcribe_async_file.py
@@ -22,7 +22,6 @@ Example usage:
 """
 
 import argparse
-import io
 
 
 # [START speech_transcribe_async]
@@ -33,7 +32,7 @@ def transcribe_file(speech_file):
     client = speech.SpeechClient()
 
     # [START speech_python_migration_async_request]
-    with io.open(speech_file, "rb") as audio_file:
+    with open(speech_file, "rb") as audio_file:
         content = audio_file.read()
 
     """
@@ -68,43 +67,10 @@ def transcribe_file(speech_file):
 # [END speech_transcribe_async]
 
 
-# [START speech_transcribe_async_gcs]
-def transcribe_gcs(gcs_uri):
-    """Asynchronously transcribes the audio file specified by the gcs_uri."""
-    from google.cloud import speech
-
-    client = speech.SpeechClient()
-
-    audio = speech.RecognitionAudio(uri=gcs_uri)
-    config = speech.RecognitionConfig(
-        encoding=speech.RecognitionConfig.AudioEncoding.FLAC,
-        sample_rate_hertz=16000,
-        language_code="en-US",
-    )
-
-    operation = client.long_running_recognize(config=config, audio=audio)
-
-    print("Waiting for operation to complete...")
-    response = operation.result(timeout=90)
-
-    # Each result is for a consecutive portion of the audio. Iterate through
-    # them to get the transcripts for the entire audio file.
-    for result in response.results:
-        # The first alternative is the most likely one for this portion.
-        print(u"Transcript: {}".format(result.alternatives[0].transcript))
-        print("Confidence: {}".format(result.alternatives[0].confidence))
-
-
-# [END speech_transcribe_async_gcs]
-
-
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(
         description=__doc__, formatter_class=argparse.RawDescriptionHelpFormatter
     )
-    parser.add_argument("path", help="File or GCS path for audio file to be recognized")
+    parser.add_argument("--path", help="File or GCS path for audio file to be recognized")
     args = parser.parse_args()
-    if args.path.startswith("gs://"):
-        transcribe_gcs(args.path)
-    else:
-        transcribe_file(args.path)
+    transcribe_file(args.path)
