+++ samples/snippets/transcribe_async_gcs_test.py
@@ -14,20 +14,13 @@
 import os
 import re
 
-import transcribe_async
+import transcribe_async_gcs
 
 RESOURCES = os.path.join(os.path.dirname(__file__), "resources")
 
 
-def test_transcribe(capsys):
-    transcribe_async.transcribe_file(os.path.join(RESOURCES, "audio.raw"))
-    out, err = capsys.readouterr()
-
-    assert re.search(r"how old is the Brooklyn Bridge", out, re.DOTALL | re.I)
-
-
 def test_transcribe_gcs(capsys):
-    transcribe_async.transcribe_gcs("gs://python-docs-samples-tests/speech/audio.flac")
+    transcribe_async_gcs.transcribe_gcs("gs://python-docs-samples-tests/speech/audio.flac")
     out, err = capsys.readouterr()
 
     assert re.search(r"how old is the Brooklyn Bridge", out, re.DOTALL | re.I)
